Correct failing test GmfGraphitiTest.openDiagramGmfAndGraphiti
The test was failing in the 2.1 series. For some reason, the old (legacy)
diagrams can no longer be adapted to some GMF class that the SWTBot code was
trying to adapt them to since Eclipse Oxygen (IDE-1853).

However, the code that was being used wasn't necessary, and wasn't really a
sufficient test for what was being tested (that appropriate editors were
opened). We now avoid the issue, and more properly test that we're seeing the
editors we expect.