IDE-2242 Fix test issue
The test was looking for the waveform, then later the component underneath it
in the Redhawk Explorer view. This has the unfortunate side-effect with SWTBot
that we've expanded the waveform, but the component may appear as a blank node
in the tree. Unless we collapse/expand the waveform, the component child's text
never gets shown.