Add appropriate delays
The test was running so fast that the notifications of model changes couldn't
make it to the UI thread in time to update the explorer. The first test would
run just fine. Then, the 2nd test would start so quickly that although the 1st
domain manager had been disposed and a 2nd one created and added, the 1st one
was still shown in the explorer view. Because the 1st domain was disposed, the
context menus for things like conn. mgr., etc. weren't available, causing the
text to fail.

The process repeated for each subsequent test (i.e. 3rd test was looking at the
2nd domain manager, which was disposed, etc).